backend/pyproject.toml
[tool.poetry]
name = "whativedone-backend"
version = "0.1.0"
description = "Backend for the What I've Done productivity tracker."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {version = "^0.30.1", extras = ["standard"]}
sqlmodel = "^0.0.19"
alembic = "^1.13.1"
psycopg = {version = "^3.1.18", extras = ["binary"]}
pydantic-settings = "^2.3.4"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
bcrypt = "^4.1.2"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"
testcontainers = "^4.6.0"
ruff = "^0.4.10"
mypy = "^1.10.0"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Configurazione degli strumenti di sviluppo
[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88
lint.select = ["E", "W", "F", "I", "C", "B"]
lint.ignore = ["E501"]